+++ packages/assistant/src/RPA/Assistant/flet_client.py
@@ -2,11 +2,11 @@ from dataclasses import dataclass
 import time
 from logging import getLogger
 from timeit import default_timer as timer
-from typing import Callable, List, Optional, Set, Tuple, Union
+from typing import Callable, Dict, List, Optional, Set, Tuple, Union
 
 import flet
 from flet import Container, Page, ScrollMode
-from flet_core import Column, Control, AppBar, Row, Stack
+from flet_core import Checkbox, Column, Control, AppBar, Row, Stack
 from flet_core.control_event import ControlEvent
 from typing_extensions import Literal
 
@@ -41,6 +41,7 @@ class Elements:
     """Lists of visible and invisible control elements"""
 
     visible: List[Control]
+    visible_by_name: Dict[str, Control]
     invisible: List[Control]
     app_bar: Optional[AppBar]
     used_names: Set[str]
@@ -55,7 +56,7 @@ class FletClient:
         self.page: Optional[Page] = None
         self.pending_operation: Optional[Callable] = None
 
-        self._elements: Elements = Elements([], [], None, set())
+        self._elements: Elements = Elements([], {}, [], None, set())
         self._to_disable: List[flet.Control] = []
         self._layout_stack: List[Union[SupportedFletLayout, AppBar]] = []
 
@@ -120,7 +121,36 @@ class FletClient:
             self.clear_elements()
             self._to_disable.clear()
 
-    def _make_flet_event_handler(self, name: str, handler: Optional[Callable] = None):
+    def _checkbox_data_saver(self, name: str, handler: Optional[Callable] = None):
+        """_make_on_change_data_saver special case for checkboxes, until
+        https://github.com/flet-dev/flet/issues/1251 is fixed"""
+
+        def handle_string_bool(bool_string: Literal["true", "false"]):
+            if bool_string == "true":
+                return True
+            elif bool_string == "false":
+                return False
+            else:
+                raise ValueError(f"Invalid checkbox value {bool_string}")
+
+        # We don't want the if inside the change_listener so it doesn't have to run on
+        # every on_change event
+        if not handler:
+
+            def change_listener(e: ControlEvent):
+                self.results[name] = handle_string_bool(e.data)
+
+        else:
+
+            def change_listener(e: ControlEvent):
+                handler(e)
+                self.results[name] = handle_string_bool(e.data)
+
+        return change_listener
+
+    def _make_on_change_data_saver(
+        self, name: str, handler: Optional[Callable] = None
+    ) -> Callable[[ControlEvent], None]:
         """Add flet event handler to record the element's data whenever content changes
         if ``handler`` is provided also call that.
         """
@@ -158,6 +188,7 @@ class FletClient:
         element: flet.Control,
         name: Optional[str] = None,
         extra_handler: Optional[Callable] = None,
+        validation_func: Optional[Callable] = None,
     ):
         if name in self._elements.used_names:
             raise ValueError(f"Name `{name}` already in use")
@@ -176,13 +207,28 @@ class FletClient:
             self._elements.visible.append(new_element)
 
         if name is not None:
+            self._elements.visible_by_name[name] = element
             self._elements.used_names.add(name)
-            element.on_change = self._make_flet_event_handler(name, extra_handler)
+
+            if isinstance(element, Checkbox):
+                on_change_data_saver = self._checkbox_data_saver(name, extra_handler)
+
+            else:
+                on_change_data_saver = self._make_on_change_data_saver(
+                    name, extra_handler
+                )
+
+            def on_change(event):
+                if validation_func:
+                    validation_func(event)
+                on_change_data_saver(event)
+
+            element.on_change = on_change
 
     def add_invisible_element(self, element: flet.Control, name: Optional[str] = None):
         self._elements.invisible.append(element)
         if name is not None:
-            element.on_change = self._make_flet_event_handler(name)
+            element.on_change = self._make_on_change_data_saver(name)
 
     def display_flet_window(
         self,
@@ -199,7 +245,7 @@ class FletClient:
         )
 
     def clear_elements(self):
-        self._elements = Elements([], [], None, set())
+        self._elements = Elements([], {}, [], None, set())
         if self.page:
             if self.page.controls:
                 self.page.controls.clear()
@@ -273,3 +319,14 @@ class FletClient:
             raise RuntimeError("Cannot determine dimensions of parent element")
 
         return (current_layout.width, current_layout.height)
+
+    def set_error(self, element_name: str, error: str):
+        """Set an error for an element. The element must be a TextField or other
+        Control that has error_text attribute."""
+        element = self._elements.visible_by_name[element_name]
+        if not hasattr(element, "error_text"):
+            raise RuntimeError(
+                "Tried to set error for element that does not have error_text attribute"
+            )
+        element.error_text = error
+        self.flet_update()